+++ sendly/__init__.py
@@ -29,7 +29,7 @@ Async Example:
     >>> asyncio.run(main())
 """
 
-__version__ = "3.32.0"
+__version__ = "3.33.0"
 
 # Main clients
 from .client import AsyncSendly, Sendly